requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
nltk==3.6.3
python-dotenv==0.19.0
regex==2021.8.28
requests==2.26.0
soupsieve==2.2.1
tika==1.24
tqdm==4.62.3
urllib3==1.26.6
Werkzeug==2.0.1
gunicorn
numpy>=1.9.2
cdifflib>= 1.2.5